requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
pytest==7.4.3
allure-pytest==2.13.2
openpyxl==3.1.2
pytest-html==4.1.1